Migrate carousel from /docs/components/carousel to /components/carousel
Port Sizes and Orientation examples from the old documentation page to
the new reference page, remove the old page file and its route, and
rewrite E2E tests to target the new URL.
