+++ site/ui/routes.tsx
@@ -37,7 +37,6 @@ import { SliderRefPage } from './pages/components/slider'
 import { ToggleGroupRefPage } from './pages/components/toggle-group'
 import { BreadcrumbPage } from './pages/breadcrumb'
 import { CalendarPage } from './pages/calendar'
-import { CarouselPage } from './pages/carousel'
 import { CardPage } from './pages/card'
 import { CheckboxPage } from './pages/checkbox'
 import { CheckboxRefPage } from './pages/components/checkbox'
@@ -481,11 +480,6 @@ export function createApp() {
 // Calendar documentation
   app.get('/docs/components/calendar', (c) => {
     return c.render(<CalendarPage />)
-  })
-
-  // Carousel documentation
-  app.get('/docs/components/carousel', (c) => {
-    return c.render(<CarouselPage />)
   })
 
   // Card documentation
